requirements.txt
fastapi
uvicorn
python-binance
python-dotenv
joblib
numpy
pandas
scikit-learn
